fix display issue

+++ include/kari/util/odometry.h
@@ -13,10 +13,11 @@ class Odom {
     int * getDL();
     int * getDR();
     int * getDM();
-    int * getLB();
-    int * getLF();
-    int * getRB();
-    int * getRF();
+
+    double * getLB();
+    double * getLF();
+    double * getRB();
+    double * getRF();
 
     double * getThetaDeg();
     double * getThetaRad();
@@ -36,7 +37,7 @@ class Odom {
 
     static int currentL, currentR, currentM;
     static int deltaL, deltaR, deltaM, lastDeltaL, lastDeltaR, lastDeltaM;
-    static int currentLF, currentLB, currentRF, currentRB;
+    static double currentLF, currentLB, currentRF, currentRB;
     static double inertL, inertR, inertT;
     static double thetaRad, thetaDeg, offset, posX, posY;
